requirements.txt
aerich==0.5.5
aiosqlite==0.16.1
asyncpg==0.23.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.68.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
iso8601==0.1.16
passlib==1.7.4
priority==2.0.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
wsproto==1.0.0
